refactor: extract useTransactionSync from App.tsx (phase 4d)
Moves the syncTransactions useEffect (~105 lines) into
src/hooks/useTransactionSync.ts. Hook takes session, authMode,
prefsLoadedRef, error/notice setters, and pref setters as parameters
(dependency injection — App.tsx stays the wiring layer).

App.tsx: 589 → 462 lines. Behavior and all 658 tests unchanged.
