+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef } from 'react'
+import { useState, useRef } from 'react'
 import { toast } from 'sonner'
 import { Toaster } from './components/ui/sonner'
 import { TatuLogo } from './components/TatuLogo'
@@ -27,25 +27,16 @@ import { Sun, Moon, Menu, Monitor } from 'lucide-react'
 import { useStore } from 'zustand'
 import { transactionStore } from './stores/transaction-store'
 import { signOut } from './services/supabase/auth'
-import { loadUserTransactions } from './services/supabase/transactions'
 import {
   clearAllCategoryOverrides,
-  replaceMerchantCategoryOverrides,
 } from './services/categorizer/category-overrides'
-import {
-  clearAllDescriptionOverrides,
-  replaceDescriptionOverrides,
-} from './services/descriptions/description-overrides'
+import { clearAllDescriptionOverrides } from './services/descriptions/description-overrides'
 import { replaceCustomCategories } from './services/categories/category-store'
 import { replaceCustomPatterns } from './services/categorizer/custom-patterns'
-import { listCategoryOverrides } from './services/supabase/category-overrides'
-import { listDescriptionOverrides as listRemoteDescriptionOverrides } from './services/supabase/description-overrides'
-import { listCustomCategories } from './services/supabase/custom-categories'
-import { listCustomPatterns as listSupabaseCustomPatterns } from './services/supabase/custom-patterns'
-import { loadUserPreferences } from './services/supabase/user-preferences'
 import { resetUserSupabaseData } from './services/supabase/reset'
 import { useUserPreferences } from './hooks/useUserPreferences'
 import { useAuthSession } from './hooks/useAuthSession'
+import { useTransactionSync } from './hooks/useTransactionSync'
 import { useTransactionHandlers } from './hooks/useTransactionHandlers'
 
 function App() {
@@ -92,112 +83,16 @@ function App() {
     setFxRate,
   } = useUserPreferences(session, prefsLoadedRef)
 
-  useEffect(() => {
-    let cancelled = false
-
-    async function syncTransactions() {
-      if (authMode === 'reset') return
-      if (!session) return
-
-      setAuthError('')
-      setAuthNotice('')
-
-      try {
-        const [
-          remoteTransactions,
-          remoteOverrides,
-          remoteDescriptionOverrides,
-          remoteCustomPatterns,
-          remoteCustomCategories,
-          userPrefs,
-        ] = await Promise.all([
-          loadUserTransactions(session),
-          listCategoryOverrides(session),
-          listRemoteDescriptionOverrides(session),
-          listSupabaseCustomPatterns(session),
-          listCustomCategories(session),
-          loadUserPreferences(session),
-        ])
-
-        replaceMerchantCategoryOverrides(
-          remoteOverrides.reduce(
-            (acc, override) => {
-              acc[override.merchantNormalized] = {
-                merchantName: override.merchantOriginal,
-                category: override.category,
-                updatedAt: override.updatedAt,
-              }
-              return acc
-            },
-            {} as Record<
-              string,
-              { merchantName?: string; category: string; updatedAt: string }
-            >
-          )
-        )
-
-        replaceDescriptionOverrides(
-          remoteDescriptionOverrides.reduce(
-            (acc, override) => {
-              acc[override.descriptionNormalized] = {
-                descriptionOriginal: override.descriptionOriginal,
-                friendlyDescription: override.friendlyDescription,
-                category: override.category,
-                updatedAt: override.updatedAt,
-              }
-              return acc
-            },
-            {} as Record<
-              string,
-              {
-                descriptionOriginal?: string
-                friendlyDescription: string
-                category?: string
-                updatedAt: string
-              }
-            >
-          )
-        )
-
-        replaceCustomPatterns(remoteCustomPatterns)
-        replaceCustomCategories(
-          remoteCustomCategories.map((category) => ({
-            id: category.id,
-            label: category.label,
-            color: category.color,
-            icon: category.icon,
-            isIgnored: category.isIgnored,
-          }))
-        )
-
-        if (userPrefs) {
-          setTheme(userPrefs.theme)
-          setPreferredCurrency(userPrefs.currency)
-          setFxRate(userPrefs.fxRate)
-        }
-        // Allow preference saves now that remote values are applied.
-        prefsLoadedRef.current = true
-
-        if (!cancelled) {
-          transactionStore.getState().setTransactions(remoteTransactions)
-        }
-      } catch (error) {
-        if (!cancelled) {
-          setAuthError(
-            error instanceof Error
-              ? error.message
-              : 'No se pudieron cargar las transacciones'
-          )
-        }
-      }
-    }
-
-    void syncTransactions()
-
-    return () => {
-      cancelled = true
-    }
-  }, [authMode, session])
+  useTransactionSync(
+    session,
+    authMode,
+    prefsLoadedRef,
+    setAuthError,
+    setAuthNotice,
+    setTheme,
+    setPreferredCurrency,
+    setFxRate,
+  )
 
   async function handleSignOut() {
     setAuthSubmitting(true)
